refactor: unificar el formato de votos en un módulo compartido
Los formateadores de votos/porcentaje estaban duplicados en el controlador de
cliente (PredictionVoteController) y en consts/predictions.ts, con riesgo de
divergencia entre el render SSR y el render en vivo.

Se extraen a src/lib/format-votes.ts (sin dependencias pesadas, para no arrastrar
battles/boxers al bundle del cliente). predictions.ts los reexporta —los
componentes los siguen importando desde @/consts/predictions sin cambios— y el
controlador los importa directamente del módulo ligero.
